Internal Machine and Material Class Refactor (#2562)
