+++ src/main/java/com/gregtechceu/gtceu/api/data/chemical/ChemicalHelper.java
@@ -11,7 +11,7 @@ import com.gregtechceu.gtceu.api.data.chemical.material.stack.UnificationEntry;
 import com.gregtechceu.gtceu.api.data.tag.TagPrefix;
 import com.gregtechceu.gtceu.api.data.tag.TagUtil;
 import com.gregtechceu.gtceu.api.fluids.store.FluidStorageKey;
-import com.gregtechceu.gtceu.common.data.GTItems;
+import com.gregtechceu.gtceu.common.data.GTMaterialItems;
 import com.gregtechceu.gtceu.data.recipe.misc.StoneMachineRecipes;
 import com.gregtechceu.gtceu.data.recipe.misc.WoodMachineRecipes;
 import com.gregtechceu.gtceu.data.tags.TagsHandler;
@@ -413,7 +413,7 @@ public class ChemicalHelper {
                 }
             });
         }
-        GTItems.toUnify.forEach(ChemicalHelper::registerUnificationItems);
+        GTMaterialItems.toUnify.forEach(ChemicalHelper::registerUnificationItems);
         WoodMachineRecipes.registerUnificationInfo();
         StoneMachineRecipes.registerUnificationInfo();
     }
