+++ src/main/java/com/gregtechceu/gtceu/common/data/GTFluids.java
@@ -37,12 +37,14 @@ public class GTFluids {
             .register();
 
     public static void init() {
+        // Register fluids for non-materials
         handleNonMaterialFluids(GTMaterials.Water, Fluids.WATER);
         handleNonMaterialFluids(GTMaterials.Lava, Fluids.LAVA);
         handleNonMaterialFluids(GTMaterials.Milk, ForgeMod.MILK);
         ForgeMod.enableMilkFluid();
-        REGISTRATE.creativeModeTab(() -> GTCreativeModeTabs.MATERIAL_FLUID);
+
         // register fluids for materials
+        REGISTRATE.creativeModeTab(() -> GTCreativeModeTabs.MATERIAL_FLUID);
         for (MaterialRegistry registry : GTCEuAPI.materialManager.getRegistries()) {
             GTRegistrate registrate = registry.getRegistrate();
             for (var material : registry.getAllMaterials()) {
