+++ src/main/java/com/gregtechceu/gtceu/common/data/GTItems.java
@@ -9,7 +9,6 @@ import com.gregtechceu.gtceu.api.data.chemical.material.MarkerMaterial;
 import com.gregtechceu.gtceu.api.data.chemical.material.MarkerMaterials;
 import com.gregtechceu.gtceu.api.data.chemical.material.Material;
 import com.gregtechceu.gtceu.api.data.chemical.material.properties.PropertyKey;
-import com.gregtechceu.gtceu.api.data.chemical.material.registry.MaterialRegistry;
 import com.gregtechceu.gtceu.api.data.chemical.material.stack.ItemMaterialInfo;
 import com.gregtechceu.gtceu.api.data.chemical.material.stack.MaterialStack;
 import com.gregtechceu.gtceu.api.data.chemical.material.stack.UnificationEntry;
@@ -18,13 +17,10 @@ import com.gregtechceu.gtceu.api.data.tag.TagUtil;
 import com.gregtechceu.gtceu.api.gui.misc.ProspectorMode;
 import com.gregtechceu.gtceu.api.item.ComponentItem;
 import com.gregtechceu.gtceu.api.item.IComponentItem;
-import com.gregtechceu.gtceu.api.item.IGTTool;
 import com.gregtechceu.gtceu.api.item.TagPrefixItem;
 import com.gregtechceu.gtceu.api.item.armor.ArmorComponentItem;
 import com.gregtechceu.gtceu.api.item.component.*;
-import com.gregtechceu.gtceu.api.item.tool.GTToolType;
 import com.gregtechceu.gtceu.api.item.tool.MaterialToolTier;
-import com.gregtechceu.gtceu.api.registry.registrate.GTRegistrate;
 import com.gregtechceu.gtceu.common.data.materials.GTFoods;
 import com.gregtechceu.gtceu.common.entity.GTBoat;
 import com.gregtechceu.gtceu.common.item.*;
@@ -64,15 +60,11 @@ import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.FluidType;
 import net.minecraftforge.fluids.FluidUtil;
 
-import com.google.common.collect.ArrayTable;
-import com.google.common.collect.ImmutableTable;
-import com.google.common.collect.Table;
 import com.tterrag.registrate.builders.ItemBuilder;
 import com.tterrag.registrate.providers.DataGenContext;
 import com.tterrag.registrate.providers.ProviderType;
 import com.tterrag.registrate.providers.RegistrateLangProvider;
 import com.tterrag.registrate.util.entry.ItemEntry;
-import com.tterrag.registrate.util.entry.ItemProviderEntry;
 import com.tterrag.registrate.util.nullness.NonNullBiConsumer;
 import com.tterrag.registrate.util.nullness.NonNullConsumer;
 import com.tterrag.registrate.util.nullness.NonNullFunction;
@@ -96,87 +88,6 @@ import static com.gregtechceu.gtceu.utils.FormattingUtil.toEnglishName;
  * @implNote GTItems
  */
 public class GTItems {
-
-    //////////////////////////////////////
-    // ***** Material Items ******//
-    //////////////////////////////////////
-
-    public static final Map<UnificationEntry, Supplier<? extends ItemLike>> toUnify = new HashMap<>();
-    public static final Map<TagPrefix, TagPrefix> purifyMap = new HashMap<>();
-
-    static {
-        purifyMap.put(TagPrefix.crushed, TagPrefix.crushedPurified);
-        purifyMap.put(TagPrefix.dustImpure, TagPrefix.dust);
-        purifyMap.put(TagPrefix.dustPure, TagPrefix.dust);
-    }
-
-    public static Table<TagPrefix, Material, ItemEntry<TagPrefixItem>> MATERIAL_ITEMS;
-
-    public static void generateMaterialItems() {
-        REGISTRATE.creativeModeTab(() -> MATERIAL_ITEM);
-        ImmutableTable.Builder<TagPrefix, Material, ItemEntry<TagPrefixItem>> builder = ImmutableTable.builder();
-        for (var tagPrefix : TagPrefix.values()) {
-            if (tagPrefix.doGenerateItem()) {
-                for (MaterialRegistry registry : GTCEuAPI.materialManager.getRegistries()) {
-                    GTRegistrate registrate = registry.getRegistrate();
-                    for (Material material : registry.getAllMaterials()) {
-                        if (tagPrefix.doGenerateItem(material)) {
-                            builder.put(tagPrefix, material, registrate
-                                    .item(tagPrefix.idPattern().formatted(material.getName()),
-                                            properties -> new TagPrefixItem(properties, tagPrefix, material))
-                                    .onRegister(TagPrefixItem::onRegister)
-                                    .setData(ProviderType.LANG, NonNullBiConsumer.noop())
-                                    .transform(unificationItem(tagPrefix, material))
-                                    .properties(p -> p.stacksTo(tagPrefix.maxStackSize()))
-                                    .model(NonNullBiConsumer.noop())
-                                    .color(() -> TagPrefixItem::tintColor)
-                                    .onRegister(GTItems::cauldronInteraction)
-                                    .register());
-                        }
-                    }
-                }
-            }
-        }
-        MATERIAL_ITEMS = builder.build();
-    }
-
-    //////////////////////////////////////
-    // ***** Material Tools ******//
-    //////////////////////////////////////
-    public final static Table<Material, GTToolType, ItemProviderEntry<IGTTool>> TOOL_ITEMS = ArrayTable.create(
-            GTCEuAPI.materialManager.getRegisteredMaterials().stream().filter(mat -> mat.hasProperty(PropertyKey.TOOL))
-                    .toList(),
-            GTToolType.getTypes().values().stream().toList());
-
-    public static void generateTools() {
-        REGISTRATE.creativeModeTab(() -> TOOL);
-
-        for (GTToolType toolType : GTToolType.getTypes().values()) {
-            for (MaterialRegistry registry : GTCEuAPI.materialManager.getRegistries()) {
-                GTRegistrate registrate = registry.getRegistrate();
-                for (Material material : registry.getAllMaterials()) {
-                    if (material.hasProperty(PropertyKey.TOOL)) {
-                        var property = material.getProperty(PropertyKey.TOOL);
-                        var tier = material.getToolTier();
-
-                        if (property.hasType(toolType)) {
-                            TOOL_ITEMS
-                                    .put(material, toolType,
-                                            (ItemProviderEntry<IGTTool>) (ItemProviderEntry<?>) registrate
-                                                    .item(toolType.idFormat.formatted(tier.material.getName()),
-                                                            p -> toolType.constructor.apply(toolType, tier, material,
-                                                                    toolType.toolDefinition, p).asItem())
-                                                    .properties(p -> p.craftRemainder(Items.AIR))
-                                                    .setData(ProviderType.LANG, NonNullBiConsumer.noop())
-                                                    .model(NonNullBiConsumer.noop())
-                                                    .color(() -> IGTTool::tintColor)
-                                                    .register());
-                        }
-                    }
-                }
-            }
-        }
-    }
 
     //////////////////////////////////////
     // ******* Misc Items ********//
@@ -565,7 +476,6 @@ public class GTItems {
             .onRegister(materialInfo(new ItemMaterialInfo(new MaterialStack(GTMaterials.Glass, GTValues.M / 4))))
             .register();
 
-    // TODO Lighter
     public static ItemEntry<ComponentItem> TOOL_MATCHES = REGISTRATE.item("matches", ComponentItem::create)
             .lang("Matches")
             .setData(ProviderType.ITEM_MODEL, NonNullBiConsumer.noop())
@@ -1351,7 +1261,6 @@ public class GTItems {
                 lines.add(Component.translatable("item.gtceu.robot.arm.tooltip"));
                 lines.add(Component.translatable("gtceu.universal.tooltip.item_transfer_rate", 8));
             })))
-            .tag(CustomTags.ROBOT_ARMS)
             .tag(CustomTags.ROBOT_ARMS)
             .register();
     public static ItemEntry<ComponentItem> ROBOT_ARM_MV = REGISTRATE.item("mv_robot_arm", ComponentItem::create)
@@ -2676,8 +2585,8 @@ public class GTItems {
             .register();
 
     public static void init() {
-        generateMaterialItems();
-        generateTools();
+        GTMaterialItems.generateMaterialItems();
+        GTMaterialItems.generateTools();
     }
 
     public static <T extends ItemLike> NonNullConsumer<T> materialInfo(ItemMaterialInfo materialInfo) {
@@ -2691,7 +2600,7 @@ public class GTItems {
             builder.onRegister(item -> {
                 Supplier<ItemLike> supplier = SupplierMemoizer.memoize(() -> item);
                 UnificationEntry entry = new UnificationEntry(tagPrefix, mat);
-                toUnify.put(entry, supplier);
+                GTMaterialItems.toUnify.put(entry, supplier);
                 ChemicalHelper.registerUnificationItems(entry, supplier);
             });
             return builder;
@@ -2699,12 +2608,13 @@ public class GTItems {
     }
 
     public static <T extends Item> void cauldronInteraction(T item) {
-        if (item instanceof TagPrefixItem tagPrefixItem && purifyMap.containsKey(tagPrefixItem.tagPrefix)) {
+        if (item instanceof TagPrefixItem tagPrefixItem &&
+                GTMaterialItems.purifyMap.containsKey(tagPrefixItem.tagPrefix)) {
             CauldronInteraction.WATER.put(item, (state, world, pos, player, hand, stack) -> {
                 if (!world.isClientSide) {
                     Item stackItem = stack.getItem();
                     if (stackItem instanceof TagPrefixItem prefixItem) {
-                        if (!purifyMap.containsKey(prefixItem.tagPrefix))
+                        if (!GTMaterialItems.purifyMap.containsKey(prefixItem.tagPrefix))
                             return InteractionResult.PASS;
                         if (!state.hasProperty(LayeredCauldronBlock.LEVEL)) {
                             return InteractionResult.PASS;
@@ -2714,8 +2624,9 @@ public class GTItems {
                         if (level == 0)
                             return InteractionResult.PASS;
 
-                        player.setItemInHand(hand, ChemicalHelper.get(purifyMap.get(prefixItem.tagPrefix),
-                                prefixItem.material, stack.getCount()));
+                        player.setItemInHand(hand,
+                                ChemicalHelper.get(GTMaterialItems.purifyMap.get(prefixItem.tagPrefix),
+                                        prefixItem.material, stack.getCount()));
                         player.awardStat(Stats.USE_CAULDRON);
                         player.awardStat(Stats.ITEM_USED.get(stackItem));
                         LayeredCauldronBlock.lowerFillLevel(state, world, pos);
