+++ src/main/java/com/gregtechceu/gtceu/common/machine/electric/RockCrusherMachine.java
@@ -2,12 +2,12 @@ package com.gregtechceu.gtceu.common.machine.electric;
 
 import com.gregtechceu.gtceu.api.machine.IMachineBlockEntity;
 import com.gregtechceu.gtceu.api.machine.SimpleTieredMachine;
-import com.gregtechceu.gtceu.common.data.GTMachines;
+import com.gregtechceu.gtceu.common.data.machines.GTMachineUtils;
 
 public class RockCrusherMachine extends SimpleTieredMachine {
 
     public RockCrusherMachine(IMachineBlockEntity holder, int tier, Object... args) {
-        super(holder, tier, GTMachines.defaultTankSizeFunction, args);
+        super(holder, tier, GTMachineUtils.defaultTankSizeFunction, args);
     }
 
     @Override
